Create H5: Tree Operations
H5: Tree Operations
Problem
Submissions
Leaderboard
A tree is an undirected graph in which any two vertices are connected by exactly one path, or equivalently a connected acyclic undirected graph. Suppose you are given a tree with n vertices in it. The vertices are numbered from 1 to n, the root of the tree is at the vertex 1.

You can perform the following operation at most k times:

choose an edge (v,u) of the tree such that v is a parent of u;
remove the edge (v,u);
add an edge (1,u) (i. e. make u with its subtree a child of the root).
Your task is to the print the minimum possible height of a given tree.

Input Format

The first line contains two integers n and k — the number of vertices in the tree and the maximum number of operations you can perform.
The second line contains n−1 integers p2,p3,…,pn — the parent of the i-th vertex. Vertex 1 is the root.
Constraints

2≤n≤2⋅10^5
0≤k≤n−1
1≤p[i] < i
Output Format

print a single integer — the smallest height of the tree that can achieved by performing at most k operations.

Sample Input 0

8 5
1 1 1 2 3 1 4
Sample Output 0

1
Sample Input 1

7 4
1 2 2 4 5 1
Sample Output 1

1