Create MT102 Game of Balls
MT102 Game of Balls
Problem
Submissions
Leaderboard
Don Snow's life is in danger. Sirsea has sentenced him to death. But Don has demanded trial by programming, and he has named you as his champion to solve the following problem :

A bag contains N balls indexed from 1 to N, ith of which has radius Ri.

One instance of a game is played as follows :

A number K is chosen.
One ball is picked out from the bag, if this ball has index K, then the game stops, otherwise the game continues.

The probability of a ball being picked from the bag is proportional to its radius.
Your task is, to find the expected number of balls to be taken out of the bag for every K from 1 to N.

INPUT
First line contains the number of balls N.
Second line contains N space separated integers, ith of which denotes Ri.

OUTPUT
Print N lines of output.
ith line must contain the expected number of balls to be taken out of the bag before the game stops if the value of K for the game is i.
Print the answer upto 6 decimal places.

CONSTRAINTS
1 ≤ N ≤ 1000
1 ≤ Ri ≤ 1000

Sample Input 0

3
1 2 3
Sample Output 0

2.4166666667
1.9333333333
1.6500000000